API: minimal endpoints for candidates CRUD
