+++ src/ApplicantTracking.Application/Commands/CreateCandidateCommand.cs
@@ -1,5 +1,7 @@
 using MediatR;
 using ApplicantTracking.Application.DTOs;
 
+namespace ApplicantTracking.Application.Commands;
+
 public record CreateCandidateCommand(string Name, string Surname, DateTime Birthdate, string Email)
     : IRequest<CandidateDto>;
