+++ src/ApplicantTracking.Application/Commands/DeleteCandidateCommand.cs
@@ -1,3 +1,5 @@
 using MediatR;
 
+namespace ApplicantTracking.Application.Commands;
+
 public record DeleteCandidateCommand(int IdCandidate) : IRequest;
