+++ src/ApplicantTracking.Application/Commands/Handlers/UpdateCandidateHandler.cs
@@ -1,10 +1,12 @@
+using ApplicantTracking.Application.Commands;
 using ApplicantTracking.Application.DTOs;
 using ApplicantTracking.Domain.Entities;
-using ApplicantTracking.Domain.Enums;
-using ApplicantTracking.Domain.Events;
+using ApplicantTracking.Domain.Enumerators;
 using ApplicantTracking.Domain.Interfaces;
 using MediatR;
 using System.Text.Json;
+
+namespace ApplicantTracking.Application.Commands.Handlers;
 
 public class UpdateCandidateHandler : IRequestHandler<UpdateCandidateCommand, CandidateDto>
 {
@@ -18,15 +20,17 @@ public class UpdateCandidateHandler : IRequestHandler<UpdateCandidateCommand, Ca
     public async Task<CandidateDto> Handle(UpdateCandidateCommand cmd, CancellationToken ct)
     {
         var current = await _repo.GetByIdAsync(cmd.IdCandidate, ct)
-                     ?? throw new KeyNotFoundException("Candidato não localizado");
+                     ?? throw new KeyNotFoundException("Candidate not found");
 
         var oldJson = JsonSerializer.Serialize(current);
         current.Update(cmd.Name, cmd.Surname, cmd.Birthdate, cmd.Email);
         _repo.Update(current);
 
-        var evt = new CandidateChangedEvent(TimelineTypes.CandidateUpdated, current.IdCandidate, null, current);
         var newJson = JsonSerializer.Serialize(current);
-        await _timeline.AddAsync(Timeline.Create((byte)evt.Type, evt.CandidateId, oldJson, newJson), ct);
+        await _timeline.AddAsync(
+            Timeline.Create((byte)TimelineTypes.CandidateUpdated, current.IdCandidate, oldJson, newJson),
+            ct
+        );
 
         await _uow.SaveChangesAsync(ct);
