+++ src/ApplicantTracking.Application/Commands/UpdateCandidateCommand.cs
@@ -1,5 +1,7 @@
 using MediatR;
 using ApplicantTracking.Application.DTOs;
 
+namespace ApplicantTracking.Application.Commands;
+
 public record UpdateCandidateCommand(int IdCandidate, string Name, string Surname, DateTime Birthdate, string Email)
     : IRequest<CandidateDto>;
